+++ TheMacroApp/SettingsForm.Designer.cs
@@ -1,6 +1,6 @@
 ﻿namespace TheMacroApp
 {
-    partial class ConfigureMacroForm
+    partial class SettingsForm
     {
         /// <summary>
         /// Required designer variable.
@@ -29,15 +29,11 @@
         private void InitializeComponent()
         {
             this.DoneButton = new System.Windows.Forms.Button();
+            this.QuitButton = new System.Windows.Forms.Button();
+            this.StartupCheckBox = new System.Windows.Forms.CheckBox();
             this.label1 = new System.Windows.Forms.Label();
-            this.ArgumentsTextBox = new System.Windows.Forms.TextBox();
-            this.TerminalComboBox = new System.Windows.Forms.ComboBox();
-            this.label2 = new System.Windows.Forms.Label();
-            this.PathTextBox = new System.Windows.Forms.TextBox();
-            this.label3 = new System.Windows.Forms.Label();
-            this.OpenFileDialogPathButton = new System.Windows.Forms.Button();
-            this.ResetMacroButton = new System.Windows.Forms.Button();
-            this.DiscardButton = new System.Windows.Forms.Button();
+            this.ScriptFolderTextBox = new System.Windows.Forms.TextBox();
+            this.ScriptFolderOpenFileDialogButton = new System.Windows.Forms.Button();
             this.HotKeyGroupBox = new System.Windows.Forms.GroupBox();
             this.HotKeyErrorLabel = new System.Windows.Forms.Label();
             this.WindowsCheckBox = new System.Windows.Forms.CheckBox();
@@ -53,103 +49,63 @@
             // DoneButton
             // 
             this.DoneButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right)));
-            this.DoneButton.Location = new System.Drawing.Point(799, 262);
+            this.DoneButton.Location = new System.Drawing.Point(551, 203);
             this.DoneButton.Name = "DoneButton";
             this.DoneButton.Size = new System.Drawing.Size(94, 29);
-            this.DoneButton.TabIndex = 0;
-            this.DoneButton.Text = "OK";
+            this.DoneButton.TabIndex = 101;
+            this.DoneButton.Text = "Done";
             this.DoneButton.UseVisualStyleBackColor = true;
             this.DoneButton.Click += new System.EventHandler(this.DoneButton_Click);
+            // 
+            // QuitButton
+            // 
+            this.QuitButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.QuitButton.Location = new System.Drawing.Point(12, 203);
+            this.QuitButton.Name = "QuitButton";
+            this.QuitButton.Size = new System.Drawing.Size(94, 29);
+            this.QuitButton.TabIndex = 100;
+            this.QuitButton.Text = "Cancel";
+            this.QuitButton.UseVisualStyleBackColor = true;
+            this.QuitButton.Click += new System.EventHandler(this.QuitButton_Click);
+            // 
+            // StartupCheckBox
+            // 
+            this.StartupCheckBox.AutoSize = true;
+            this.StartupCheckBox.Location = new System.Drawing.Point(12, 12);
+            this.StartupCheckBox.Name = "StartupCheckBox";
+            this.StartupCheckBox.Size = new System.Drawing.Size(167, 24);
+            this.StartupCheckBox.TabIndex = 0;
+            this.StartupCheckBox.Text = "Add to Startup Apps";
+            this.StartupCheckBox.UseVisualStyleBackColor = true;
             // 
             // label1
             // 
             this.label1.AutoSize = true;
-            this.label1.Location = new System.Drawing.Point(12, 48);
+            this.label1.Location = new System.Drawing.Point(12, 45);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(84, 20);
-            this.label1.TabIndex = 1;
-            this.label1.Text = "Arguments:";
+            this.label1.Size = new System.Drawing.Size(149, 20);
+            this.label1.TabIndex = 102;
+            this.label1.Text = "Default Script Folder:";
             // 
-            // ArgumentsTextBox
+            // ScriptFolderTextBox
             // 
-            this.ArgumentsTextBox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
+            this.ScriptFolderTextBox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
-            this.ArgumentsTextBox.Location = new System.Drawing.Point(124, 45);
-            this.ArgumentsTextBox.Name = "ArgumentsTextBox";
-            this.ArgumentsTextBox.Size = new System.Drawing.Size(769, 27);
-            this.ArgumentsTextBox.TabIndex = 2;
+            this.ScriptFolderTextBox.Location = new System.Drawing.Point(186, 42);
+            this.ScriptFolderTextBox.Name = "ScriptFolderTextBox";
+            this.ScriptFolderTextBox.Size = new System.Drawing.Size(429, 27);
+            this.ScriptFolderTextBox.TabIndex = 103;
             // 
-            // TerminalComboBox
+            // ScriptFolderOpenFileDialogButton
             // 
-            this.TerminalComboBox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.TerminalComboBox.FormattingEnabled = true;
-            this.TerminalComboBox.Location = new System.Drawing.Point(124, 78);
-            this.TerminalComboBox.Name = "TerminalComboBox";
-            this.TerminalComboBox.Size = new System.Drawing.Size(769, 28);
-            this.TerminalComboBox.TabIndex = 3;
-            // 
-            // label2
-            // 
-            this.label2.AutoSize = true;
-            this.label2.Location = new System.Drawing.Point(12, 81);
-            this.label2.Name = "label2";
-            this.label2.Size = new System.Drawing.Size(69, 20);
-            this.label2.TabIndex = 4;
-            this.label2.Text = "Terminal:";
-            // 
-            // PathTextBox
-            // 
-            this.PathTextBox.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.PathTextBox.Location = new System.Drawing.Point(124, 12);
-            this.PathTextBox.Name = "PathTextBox";
-            this.PathTextBox.Size = new System.Drawing.Size(736, 27);
-            this.PathTextBox.TabIndex = 6;
-            this.PathTextBox.TextChanged += new System.EventHandler(this.PathTextBox_TextChanged);
-            // 
-            // label3
-            // 
-            this.label3.AutoSize = true;
-            this.label3.Location = new System.Drawing.Point(12, 15);
-            this.label3.Name = "label3";
-            this.label3.Size = new System.Drawing.Size(40, 20);
-            this.label3.TabIndex = 5;
-            this.label3.Text = "Path:";
-            // 
-            // OpenFileDialogPathButton
-            // 
-            this.OpenFileDialogPathButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
-            this.OpenFileDialogPathButton.Location = new System.Drawing.Point(866, 12);
-            this.OpenFileDialogPathButton.Name = "OpenFileDialogPathButton";
-            this.OpenFileDialogPathButton.Size = new System.Drawing.Size(27, 27);
-            this.OpenFileDialogPathButton.TabIndex = 7;
-            this.OpenFileDialogPathButton.Text = "...";
-            this.OpenFileDialogPathButton.UseVisualStyleBackColor = true;
-            this.OpenFileDialogPathButton.Click += new System.EventHandler(this.OpenFileDialogPathButton_Click);
-            // 
-            // ResetMacroButton
-            // 
-            this.ResetMacroButton.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Left) 
-            | System.Windows.Forms.AnchorStyles.Right)));
-            this.ResetMacroButton.Location = new System.Drawing.Point(124, 225);
-            this.ResetMacroButton.Name = "ResetMacroButton";
-            this.ResetMacroButton.Size = new System.Drawing.Size(769, 29);
-            this.ResetMacroButton.TabIndex = 8;
-            this.ResetMacroButton.Text = "Reset";
-            this.ResetMacroButton.UseVisualStyleBackColor = true;
-            this.ResetMacroButton.Click += new System.EventHandler(this.ResetMacroButton_Click);
-            // 
-            // DiscardButton
-            // 
-            this.DiscardButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
-            this.DiscardButton.Location = new System.Drawing.Point(12, 262);
-            this.DiscardButton.Name = "DiscardButton";
-            this.DiscardButton.Size = new System.Drawing.Size(94, 29);
-            this.DiscardButton.TabIndex = 9;
-            this.DiscardButton.Text = "Cancel";
-            this.DiscardButton.UseVisualStyleBackColor = true;
-            this.DiscardButton.Click += new System.EventHandler(this.DiscardButton_Click);
+            this.ScriptFolderOpenFileDialogButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.ScriptFolderOpenFileDialogButton.Location = new System.Drawing.Point(621, 42);
+            this.ScriptFolderOpenFileDialogButton.Name = "ScriptFolderOpenFileDialogButton";
+            this.ScriptFolderOpenFileDialogButton.Size = new System.Drawing.Size(27, 27);
+            this.ScriptFolderOpenFileDialogButton.TabIndex = 104;
+            this.ScriptFolderOpenFileDialogButton.Text = "...";
+            this.ScriptFolderOpenFileDialogButton.UseVisualStyleBackColor = true;
+            this.ScriptFolderOpenFileDialogButton.Click += new System.EventHandler(this.ScriptFolderOpenFileDialogButton_Click);
             // 
             // HotKeyGroupBox
             // 
@@ -163,21 +119,21 @@
             this.HotKeyGroupBox.Controls.Add(this.label5);
             this.HotKeyGroupBox.Controls.Add(this.label4);
             this.HotKeyGroupBox.Controls.Add(this.KeyComboBox);
-            this.HotKeyGroupBox.Location = new System.Drawing.Point(124, 112);
+            this.HotKeyGroupBox.Location = new System.Drawing.Point(12, 75);
             this.HotKeyGroupBox.Name = "HotKeyGroupBox";
-            this.HotKeyGroupBox.Size = new System.Drawing.Size(769, 107);
-            this.HotKeyGroupBox.TabIndex = 10;
+            this.HotKeyGroupBox.Size = new System.Drawing.Size(633, 121);
+            this.HotKeyGroupBox.TabIndex = 105;
             this.HotKeyGroupBox.TabStop = false;
-            this.HotKeyGroupBox.Text = "Hot Key";
+            this.HotKeyGroupBox.Text = "Application Hot Key";
             // 
             // HotKeyErrorLabel
             // 
             this.HotKeyErrorLabel.Anchor = ((System.Windows.Forms.AnchorStyles)(((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left) 
             | System.Windows.Forms.AnchorStyles.Right)));
             this.HotKeyErrorLabel.BackColor = System.Drawing.Color.Lime;
-            this.HotKeyErrorLabel.Location = new System.Drawing.Point(6, 84);
+            this.HotKeyErrorLabel.Location = new System.Drawing.Point(6, 88);
             this.HotKeyErrorLabel.Name = "HotKeyErrorLabel";
-            this.HotKeyErrorLabel.Size = new System.Drawing.Size(757, 20);
+            this.HotKeyErrorLabel.Size = new System.Drawing.Size(615, 20);
             this.HotKeyErrorLabel.TabIndex = 7;
             this.HotKeyErrorLabel.Text = "Hot Key is VALID.";
             this.HotKeyErrorLabel.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
@@ -251,36 +207,32 @@
             this.KeyComboBox.FormattingEnabled = true;
             this.KeyComboBox.Location = new System.Drawing.Point(103, 26);
             this.KeyComboBox.Name = "KeyComboBox";
-            this.KeyComboBox.Size = new System.Drawing.Size(660, 28);
+            this.KeyComboBox.Size = new System.Drawing.Size(518, 28);
             this.KeyComboBox.TabIndex = 0;
-            this.KeyComboBox.SelectedIndexChanged += new System.EventHandler(this.OnKeyChange);
+            this.KeyComboBox.SelectedValueChanged += new System.EventHandler(this.OnKeyChange);
             // 
-            // ConfigureMacroForm
+            // SettingsForm
             // 
             this.AcceptButton = this.DoneButton;
             this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 20F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.CancelButton = this.DiscardButton;
-            this.ClientSize = new System.Drawing.Size(905, 303);
+            this.CancelButton = this.QuitButton;
+            this.ClientSize = new System.Drawing.Size(657, 244);
             this.Controls.Add(this.HotKeyGroupBox);
-            this.Controls.Add(this.DiscardButton);
-            this.Controls.Add(this.ResetMacroButton);
-            this.Controls.Add(this.OpenFileDialogPathButton);
-            this.Controls.Add(this.PathTextBox);
-            this.Controls.Add(this.label3);
-            this.Controls.Add(this.label2);
-            this.Controls.Add(this.TerminalComboBox);
-            this.Controls.Add(this.ArgumentsTextBox);
+            this.Controls.Add(this.ScriptFolderOpenFileDialogButton);
+            this.Controls.Add(this.ScriptFolderTextBox);
             this.Controls.Add(this.label1);
+            this.Controls.Add(this.StartupCheckBox);
+            this.Controls.Add(this.QuitButton);
             this.Controls.Add(this.DoneButton);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.FixedSingle;
             this.MaximizeBox = false;
             this.MinimizeBox = false;
-            this.MinimumSize = new System.Drawing.Size(246, 350);
-            this.Name = "ConfigureMacroForm";
+            this.Name = "SettingsForm";
             this.ShowIcon = false;
             this.ShowInTaskbar = false;
-            this.Text = "Configure Macro";
-            this.Load += new System.EventHandler(this.ConfigureMacroForm_Load);
+            this.Text = "Settings";
+            this.Load += new System.EventHandler(this.SettingsForm_Load);
             this.HotKeyGroupBox.ResumeLayout(false);
             this.HotKeyGroupBox.PerformLayout();
             this.ResumeLayout(false);
@@ -291,23 +243,19 @@
         #endregion
 
         private Button DoneButton;
+        private Button QuitButton;
+        private CheckBox StartupCheckBox;
         private Label label1;
-        private TextBox ArgumentsTextBox;
-        private ComboBox TerminalComboBox;
-        private Label label2;
-        private TextBox PathTextBox;
-        private Label label3;
-        private Button OpenFileDialogPathButton;
-        private Button ResetMacroButton;
-        private Button DiscardButton;
+        private TextBox ScriptFolderTextBox;
+        private Button ScriptFolderOpenFileDialogButton;
         private GroupBox HotKeyGroupBox;
-        private Label label4;
-        private ComboBox KeyComboBox;
-        private CheckBox AltCheckBox;
-        private Label label5;
+        private Label HotKeyErrorLabel;
         private CheckBox WindowsCheckBox;
         private CheckBox ShiftCheckBox;
         private CheckBox CtrlCheckBox;
-        private Label HotKeyErrorLabel;
+        private CheckBox AltCheckBox;
+        private Label label5;
+        private Label label4;
+        private ComboBox KeyComboBox;
     }
 }